+++ package.json
@@ -13,5 +13,11 @@
   "engines": {
     "node": ">= 16.0"
   },
-  "packageManager": "yarn@4.3.1"
+  "scripts": {
+    "build": "tsc"
+  },
+  "packageManager": "yarn@4.3.1",
+  "devDependencies": {
+    "typescript": "^5.5.4"
+  }
 }
